scripts/requirements.txt
# Python-Abhängigkeiten für die Datenextraktion

# MCP Server Abfragen
requests>=2.28.0
pyyaml>=6.0

# LSN Datenbank Scraping
playwright>=1.40.0

# Datenverarbeitung
pandas>=2.0.0
openpyxl>=3.1.0  # Für Excel-Export
